Revert the undo withdrawal: it deleted other devices' sessions (KAN-83)
KAN-80 made undo tombstone every session missing from the restored
snapshot, on the assumption that the only such session is the one the
undo retracts. That assumption is false, and the consequence is data
loss: a session created on ANOTHER device is tombstoned by an undo here,
and the tombstone propagates and deletes it there too.

The undo history and the merged container are two independent notions of
"what this device has", and they drift apart at different rates.
replaceState -- what the merge uses -- is not in actionsToCapture, so a
session arriving from the cloud enters the container but no snapshot. And
setPresentStartup refreshes `present` after the first sync while never
refreshing `past`. So no comparison available inside that reducer
reliably means "the user retracted this": diffing the snapshot against
live state withdraws cloud arrivals, and diffing the two history entries
withdraws them too, because only one of the pair is ever refreshed.

Two attempts at scoping the withdrawal failed for that reason, which is
the signal to stop patching. The mismatch is architectural, and a third
guess at which pair of stale snapshots happens to be safe would be
exactly that -- a guess. So this reverts #180 whole: KAN-80 and KAN-81
both come out, and both are reopened.

The trade is not close. Without them, undoing a create bounces back with
a false "synced from another device" toast, which is annoying and
cosmetic. With them, undo silently deletes sessions from other devices,
which is unbounded and unrecoverable.

undoDoesNotDeleteRemoteSessions.test.ts is the guard that keeps this from
coming back. It brings a session in from the cloud, undoes a local save,
and asserts the remote session is not tombstoned and survives the next
merge. It asserts the TOMBSTONE rather than the session's momentary
absence: a session vanishing locally is harmless, because anything with
no tombstone against it is restored by the next merge. Re-applying #180
fails it.
